+++ src/anemoi/inference/outputs/gribfile.py
@@ -90,6 +90,8 @@ class GribFileOutput(GribOutput):
         grib1_keys=None,
         grib2_keys=None,
         modifiers=None,
+        output_frequency=None,
+        write_initial_state=None,
         **kwargs,
     ):
         super().__init__(
@@ -99,6 +101,8 @@ class GribFileOutput(GribOutput):
             grib1_keys=grib1_keys,
             grib2_keys=grib2_keys,
             modifiers=modifiers,
+            output_frequency=output_frequency,
+            write_initial_state=write_initial_state,
         )
         self.path = path
         self.output = ekd.new_grib_output(self.path, split_output=True, **kwargs)