+++ src/anemoi/inference/runner.py
@@ -64,6 +64,8 @@ class Runner(Context):
         inference_options=None,
         patch_metadata={},
         development_hacks={},  # For testing purposes, don't use in production
+        output_frequency=None,
+        write_initial_state=True,
     ):
         self._checkpoint = Checkpoint(checkpoint, patch_metadata=patch_metadata)
 
@@ -77,6 +79,8 @@ class Runner(Context):
         self.use_grib_paramid = use_grib_paramid
         self.development_hacks = development_hacks
         self.hacks = bool(development_hacks)
+        self.output_frequency = output_frequency
+        self.write_initial_state = write_initial_state
 
         # This could also be passed as an argument